Using byte (i8) to save some space on Graph

+++ src/main/c/graph.c
@@ -4,7 +4,7 @@
 Graph graph_new(i32 edges){
   Graph graph = malloc(sizeof(struct Graph));
   graph->edges_size = edges;
-  graph->edges = calloc(edges * edges, sizeof(i32));
+  graph->edges = calloc(edges * edges, sizeof(i8));
   return graph;
 }
 
@@ -17,12 +17,12 @@ size_t calc_offset(Graph graph, i32 x, i32 y){
   return x + graph->edges_size * y;
 }
 
-void graph_add_edge(Graph graph, i32 from, i32 to, i32 weight){
+void graph_add_edge(Graph graph, i32 from, i32 to, i8 weight){
   graph->edges[calc_offset(graph, from, to)] = weight;
   graph->edges[calc_offset(graph, to, from)] = weight;
 }
 
-i32 graph_get_weight(Graph graph, i32 from, i32 to){
+i8 graph_get_weight(Graph graph, i32 from, i32 to){
   return graph->edges[calc_offset(graph, from, to)];
 }
 
@@ -30,7 +30,7 @@ List get_neighbors(Graph this, i32 vertex){
   List neighbors = list_new(this->edges_size / 2);
   for(i32 i = 0; i< this->edges_size; i ++){
     if(vertex != i){
-      i32 weight = graph_get_weight(this, i, vertex);
+      i8 weight = graph_get_weight(this, i, vertex);
       if(weight > 0){
         list_add(neighbors, i);
       }
@@ -40,7 +40,7 @@ List get_neighbors(Graph this, i32 vertex){
 }
 
 i32* calculate_cost_table(Graph this, i32 from){
-  i32* costs = malloc(this->edges_size * sizeof(i32));
+  i8* costs = malloc(this->edges_size * sizeof(i8));
   i32* source = calloc(this->edges_size, sizeof(i32));
   for(i32 i = 0; i < this->edges_size;i++){
     costs[i] = source[i] = -1;    
@@ -55,12 +55,12 @@ i32* calculate_cost_table(Graph this, i32 from){
 
   while(!priority_is_empty(queue)){
     i32 current = priority_queue_pop(queue);
-    i32 currentCost = costs[current];
+    i8 currentCost = costs[current];
 
     List neighbors = get_neighbors(this, current);
     for(i32 i = 0; i < neighbors->size; i++){
       i32 neighbor = list_get(neighbors, i);
-      i32 cost = currentCost + graph_get_weight(this, neighbor, current);
+      i8 cost = currentCost + graph_get_weight(this, neighbor, current);
       i32 neighbor_cost = costs[neighbor];
 
       if(neighbor_cost == -1 || neighbor_cost > cost){
